only keep data from agrimoney and euractiv
The reason for this being that we would like to go all the way back to 2010 where there has been
significant upward trend. Neither world grain nor Bloomberg goes as far back to 2010 and is
thus removed. We would test the model first, then sequentially add the other two sources
back and assess the impact