commented out spider logging
The logging is the main reason why the scraper fails when it is executed by Airflow. The
titles are non-unicode objects which when attempted to write to file it causes an
exception. This is not the case for stand alone execution as the logs for Scrapy
has been disabled in the settings file.