fixed commodity tagging
The commodity tagging has not been functioning since the last integration. After
a thorough investigation, it was found that this is due to the change in input.
Previously, the articles were a list of tokens but it has now been changed to a single
string. We split the string into tokens to resolve this issue