reconciled the difference between the implementation between R and Python
There were minor differences between the implementation, the difference was actually
due to the default setting of noramlisation of the response. It is defaulted to True in R,
while False for Python. Normalisation usually improves the fit and prediction, this is
also supported by the examination of the prediction. Thus, we have adjusted both to
standardise the response